Refactor robot entry point and add structured logging
