+++ Logging/LogLevel.cs
@@ -1,0 +1,11 @@
+namespace LandocsRobot.Logging
+{
+    internal enum LogLevel
+    {
+        Fatal = 1,
+        Error = 2,
+        Warning = 3,
+        Info = 4,
+        Debug = 5
+    }
+}
